update Camp Management component
